Fix bug of mode operation
Former-commit-id: 30d48a17f3ac43a82d9d24ea93b94be468306e35

+++ src/bytecode/ByteCodeOperations.cpp
@@ -146,7 +146,7 @@ NEVER_INLINE ESValue modOperation(ESValue* left, ESValue* right)
             bool isLNeg = lvalue < 0.0;
             lvalue = std::abs(lvalue);
             rvalue = std::abs(rvalue);
-            int d = lvalue / rvalue;
+            uint64_t d = lvalue / rvalue;
             double r = lvalue - (d * rvalue);
             if (isLNeg)
                 r = -r;
